% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daysfromtreatment.R
\name{DaysFromTreatment}
\alias{DaysFromTreatment}
\title{Predict weather a sample has been IFN treated or not ONLY USE IF YOUR ARE SURE THAT THESE SAMPLES HAVE BEEN TREATED WITH IFNbeta}
\usage{
DaysFromTreatment(input_df)
}
\arguments{
\item{input_df}{A matrix or Dataframe of beta values}
}
\value{
A named vector containing predicted days since first treatment.
}
\description{
Predict weather a sample has been IFN treated or not ONLY USE IF YOUR ARE SURE THAT THESE SAMPLES HAVE BEEN TREATED WITH IFNbeta
}
\examples{
restults <- DaysFromTreatment(df)
restults <- DaysFromTreatment(mat)
}
