% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_Data_Matrix.R
\docType{data}
\name{Example_Data_Matrix}
\alias{Example_Data_Matrix}
\title{Example Beta Matrix from the IFNMeth package}
\format{
An object of class RDa
}
\usage{
data(Example_Data_Matrix)
}
\description{
Example Beta Matrix from the IFNMeth package
}
\examples{
data(Example_Data_Matrix)
}
\keyword{datasets}
