% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreated.R
\name{get_score}
\alias{get_score}
\title{Internal function to process beta-value columns into methylation treatment scores}
\usage{
get_score(sample_beta_column)
}
\description{
Internal function to process beta-value columns into methylation treatment scores
}
