% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifn_reference_cpgs.R
\docType{data}
\name{ifn_reference_cpgs}
\alias{ifn_reference_cpgs}
\title{Reference weights for CpGs used to create IFN methylation treatment score}
\format{
An object of class RDa
}
\usage{
data(ifn_reference_cpgs)
}
\description{
Reference weights for CpGs used to create IFN methylation treatment score
}
\examples{
data(ifn_reference_cpgs)
}
\keyword{datasets}
