% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreated.R
\name{isTreated}
\alias{isTreated}
\title{Predict weather a sample has been IFN treated or not}
\usage{
isTreated(input_df, threshold = -0.42)
}
\arguments{
\item{input_df}{A matrix or Dataframe of beta values}

\item{threshold}{Threshold for methylation score (double), -0.42 as default}
}
\value{
A dataframe with two columns,all_scores and ifn_status
}
\description{
Predict weather a sample has been IFN treated or not
}
\examples{
restults <- isTreated(df, -0.3)
restults <- isTreated(mat)
}
